# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth controller",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Creates a new user account",
                "operationId": "6703d7156c265a1e7b47e500d2a16304",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-05T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "description": "Authenticate user and return an access token",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid login credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/who-am-i": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get currently authenticated user",
                "description": "Returns the session related user",
                "operationId": "a4c83ca7601ae145633ec71a1ff848d0",
                "responses": {
                    "200": {
                        "description": "Authenticated user returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No user is authenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get list of all restaurants",
                "operationId": "fd30e9e4b362b51448c8ba783f6c93ad",
                "responses": {
                    "200": {
                        "description": "List of restaurants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Restaurant"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurants/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get restaurant by ID with menus",
                "operationId": "dd2f38814da0909c0475203a72c03f8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of restaurant to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant details with menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantWithMenus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                }
            }
        },
        "/api/restaurants/{id}/menu": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get menus of a specific restaurant by ID",
                "operationId": "a51372a83142c2a16f48b80c19e9948e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of menus for the restaurant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Menu"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Menu": {
                "title": "Menu",
                "required": [
                    "id",
                    "name",
                    "restaurant_id"
                ],
                "properties": {
                    "id": {
                        "description": "App\\Models\\Menu",
                        "type": "integer",
                        "example": 1
                    },
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lunch Menu"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Restaurant": {
                "title": "Restaurant",
                "description": "Restaurant model",
                "required": [
                    "id",
                    "name",
                    "address"
                ],
                "properties": {
                    "id": {
                        "description": "App\\Models\\Restaurant",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pizza House"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Berlin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "RestaurantWithMenus": {
                "title": "Restaurant with menus",
                "required": [
                    "id",
                    "name",
                    "address",
                    "menus"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pizza House"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Berlin"
                    },
                    "menus": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Menu"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Restaurants",
            "description": "Restaurants"
        }
    ]
}